Improve handling of ambiguous stations
If we find more than one station let's first check if there's one that
matches the given name exactly, and if so use it; the MVG API sometimes
guesses overly much.

If we don't find a station which matches exactly show the names of all
stations that MVG returned, to aid the user in picking the proper
station.